+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "spanner_emulator_toolkit"
+require "debug"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -11,5 +12,17 @@ RSpec.configure do |config|
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
+  end
+
+  config.before(:suite) do
+    SpannerEmulatorToolkit.configure do |config|
+      config.project_id = "test-project"
+      config.instance_id = "test-instance"
+      config.database_id = "test-database-#{Time.now.to_i}"
+      config.emulator_host = "localhost:9010"
+
+      config.logger = Logger.new("/dev/null")
+      config.log_level = Logger::FATAL
+    end
   end
 end